refactored to config package
